==> notebooks/ground-truth/5-AP-distribution.ipynb <==
# 5) Validation - distribution of GT and optimized AP potentials

This notebook evaluates the distributions of membrane potentials of an action potential over the neuron morphology.

import pickle
import pandas as pd
import seaborn as sns
import sys
import shutil

import bluepyopt as bpopt
import bluepyopt.ephys as ephys

import matplotlib.pyplot as plt
from scipy.spatial import distance
import MEAutility as mu
import json
import time
import numpy as np
from pathlib import Path

import pandas as pd
import seaborn as sns
from scipy.spatial.distance import cosine

import multimodalfitting as mf

%matplotlib notebook

base_dir = Path("../..")

save_fig = True
figure_folder = Path("..") / "figures"

if save_fig:
    figure_folder.mkdir(exist_ok=True)

colors_dict = {"soma": "C0",
               "all": "C1",
               "sections": "C2",
               "single": "C3"}
figsize = (10, 7)

## Load results and define model

probe_type = "planar"
model_name = "hay_ais" # "hay", "hay_ais", "hay_ais_hillock"
cell_models_folder = base_path / "cell_models"

model_folder = cell_models_folder / f"{model_name}"

result_dates = [r.name for r in (base_dir / "results").iterdir()]
# use latest results
results_date = max(result_dates)
result_folder = base_dir / "results" / results_date

pkl_file_name = "runs.pkl"

data = pickle.load(open(result_folder / pkl_file_name, 'rb'))
df_optimization = pd.DataFrame(data)
df_model = df_optimization.query(f"model == '{model_name}'")

if (result_folder / f"all_responses_{model_name}.pkl").is_file():
    with open(result_folder / f"all_responses_{model_name}.pkl", 'rb') as f:
        opt_results_all = pickle.load(f)

protocols_file = model_folder / "fitting" / "efeatures" / "protocols_BPO_all.json"
features_file = model_folder / "fitting" / "efeatures" / "features_BPO_all.json"

cell = mf.create_ground_truth_model(model_name=model_name, release=False)
cell_release = mf.create_ground_truth_model(model_name=model_name, release=True)

probe = mf.define_electrode(probe_type=probe_type)

param_names = [param.name for param in cell.params.values() if not param.frozen]

params_release = {}
for param in cell_release.params_by_names(param_names):
    params_release[param.name] = param.value

# protocol_for_eap = "IDrest_300"
protocol_for_eap = "firepattern_120"

### Define more recording points 

extra_kwargs = mf.utils.get_extra_kwargs()

eva_extra = mf.create_evaluator(
    model_name=model_name,
    feature_set="extra",
    strategy="all",
    all_protocols=True,
    protocols_with_lfp=protocol_for_eap,
    **extra_kwargs
)

mf.plot_cell(eva_extra.cell_model, eva_extra.sim, param_values=params_release, color_ais="g")

positions = np.array([[-62, 828], [-3, 954], [-27, 546], [-27, 85], [134, -28], [-117, -189],
                     [13.4, -28.7], [11.7, -1.2]]) #[12.5, -15.6],
position_names = ["apical_distal_left", "apical_distal_right", "apical_middle", "apical_proximal", 
                  "basal_right", "basal_left", "ais_distal", "ais_proximal"] #"ais_middle", 

extra_recordings = mf.utils.extra_recordings_from_positions(cell_release, eva_extra.sim, positions, position_names)

extra_recordings

eva_extra = mf.create_evaluator(
    model_name=model_name,
    feature_set="extra",
    strategy="all",
    protocols_with_lfp=protocol_for_eap,
    all_protocols=True,
    extra_recordings={protocol_for_eap: extra_recordings}, 
    **extra_kwargs
)

# Load protocols and original features

protocol_to_run = eva_extra.fitness_protocols[protocol_for_eap]
print(protocol_to_run)

strategies = ['soma', 'all', 'sections', 'single']

print("Computing release")
responses_release = eva_extra.run_protocol(protocol_to_run, params_release)

for strategy in strategies:
    print(f"Computing '{strategy}' -- seed: {opt_results_all[strategy]['best_seed']}")
    best_params = opt_results_all[strategy]["best_params"]
    responses = eva_extra.run_protocol(protocol_to_run, best_params)
    opt_results_all[strategy]["responses"] = responses

mf.plot_responses(opt_results_all["all"]["responses"], protocol_names=["soma", "ais"])

ms_after = 50
ms_before = 10

responses_cut_release = mf.utils.get_peak_cutout(responses_release, ms_before=ms_before, 
                                                 ms_after=ms_after, average=True)

for strategy in strategies:
    print(f"Cutting {strategy}")
    responses = opt_results_all[strategy]["responses"]
    responses_cut = mf.utils.get_peak_cutout(responses, ms_before=ms_before, 
                                             ms_after=ms_after, average=True)
    opt_results_all[strategy]["responses_cut"] = responses_cut

responses_to_plot = []
for resp in responses_release:
    if any(pos_name in resp for pos_name in position_names):
        responses_to_plot.append(resp)

xlims = {"apical_distal_left": [-ms_before, ms_after], 
         "apical_distal_right": [-ms_before, ms_after],
         "apical_middle": [-ms_before, ms_after], 
         "apical_proximal": [-5, 20], 
         "basal_right": [-3, 10],
         "basal_left": [-3, 10],
         "ais_distal": [-2, 4],
         "ais_middle": [-2, 4], 
         "ais_proximal": [-2, 4]}

figs_traces = {}

exclude_strategies = []

add_title = False

figs_dists = {}

order = ["soma", "all", "sections", "single"]

def compute_features(times, voltage, pre_median):
    pre_samples = np.searchsorted(times, pre_median)
    baseline = np.median(voltage[pre_samples[0]:pre_samples[1]])
    max_idx = np.argmax(voltage)
    amplitude = np.max(voltage) - baseline
    decay_time = times[max_idx + np.where(voltage[max_idx:] <= baseline)[0][0]] - times[max_idx]
    max_time = times[max_idx]
    amp50 = 0.5 * amplitude + baseline
    amp50b = np.where(voltage >= amp50)[0][0]
    amp50e = np.where(voltage >= amp50)[0][-1]
    t50b = times[amp50b]
    t50e = times[amp50e]
    fwhm = times[amp50e] - times[amp50b]
    voltage_after_decay = voltage[np.where(times >= max_time + decay_time)]
    ahp_v = np.abs(np.min(voltage_after_decay) - baseline)
    
    return amplitude, fwhm, decay_time, ahp_v 

pre_median = [-5, -1]

distance_arr = []
amp_arr = []
fwhm_arr = []
decay_arr = []
ahp_arr = []
total_score_arr = []

strategy_arr = []
position_arr = []
for resp_name in responses_to_plot:
    fig, ax = plt.subplots(figsize=figsize)
    position = resp_name.split(".")[1]
    times = np.linspace(-ms_before, ms_after, len(responses_cut_release[resp_name]["time"]))
    ax.plot(times, responses_cut_release[resp_name]["voltage"], 
            color="k", label="GT", lw=3)
    amp_gt, fwhm_gt, decay_time_gt, ahp_gt = compute_features(times,
                                                                responses_cut_release[resp_name]["voltage"], 
                                                                pre_median)
    for strategy in strategies:
        if strategy not in exclude_strategies:
            resp_cut = opt_results_all[strategy]["responses_cut"]
            ax.plot(times, resp_cut[resp_name]["voltage"], 
                    color=colors_dict[strategy], label=strategy.upper(), alpha=0.7, lw=1.5)
            dist = np.sum(np.abs(responses_cut_release[resp_name]["voltage"] - resp_cut[resp_name]["voltage"]))
            position_arr.append(position)
            distance_arr.append(dist)
            strategy_arr.append(strategy)
            
            amp, fwhm, decay_time, ahp = compute_features(times,
                                                          resp_cut[resp_name]["voltage"], 
                                                          pre_median)
            amp_score = np.abs(amp_gt - amp) / np.abs(amp_gt)
            fwhm_score = np.abs(fwhm_gt - fwhm) / np.abs(fwhm_gt)
            decay_score = np.abs(decay_time - decay_time_gt) / np.abs(decay_time_gt)
            ahp_score = np.abs(ahp_gt - ahp) / np.abs(ahp)
            
            amp_arr.append(amp_score)
            fwhm_arr.append(fwhm_score)
            decay_arr.append(decay_score)
            ahp_arr.append(ahp_score)
            
            total_score_arr.append((amp_score + fwhm_score + decay_score + ahp_score) / 4)
            

    ax.set_xlabel("time (ms)", fontsize=30)
    ax.set_ylabel("$V_m$ (mV)", fontsize=30)
    ax.set_xlim(xlims[position])
    ax.tick_params(axis='x', labelsize=20)
    ax.tick_params(axis='y', labelsize=20)
    ax.spines["top"].set_visible(False)
    ax.spines["right"].set_visible(False)
    if add_title:
        ax.set_title(f"{position}", fontsize=15)
    ax.axvline(0, color="gray", ls="--")
    ax.legend(fontsize=22)
    figs_traces[position] = fig

df = pd.DataFrame({"strategy": strategy_arr, "distance": distance_arr, "position": position_arr,
                   "amp": amp_arr, "fwhm": fwhm_arr, "decay": decay_arr, "ahp": ahp_arr,
                   "total_error": total_score_arr})

add_title = False
for position in np.unique(df["position"]):
    fig, ax = plt.subplots(figsize=figsize)
    df_pos = df.query(f"position == '{position}'")
    sns.barplot(data=df_pos, x="strategy", y="total_error", ax=ax, order=order)
    
    ax.set_xlabel("", fontsize=30)
    ax.set_ylabel("Mean relative error", fontsize=30)
    

    ax.tick_params(axis='x', labelsize=25)
    ax.tick_params(axis='y', labelsize=20)
    ax.spines["top"].set_visible(False)
    ax.spines["right"].set_visible(False)
    if add_title:
        ax.set_title(f"{position}", fontsize=15)
    figs_dists[position] = fig

fig_cell, ax_cell = plt.subplots(figsize=(7, 10))
mf.plot_cell(eva_extra.cell_model, eva_extra.sim, param_values=params_release, ax=ax_cell,
             alpha=0.5, color="black", detailed=True)

cmap = "tab20"
cm = plt.get_cmap(cmap)
for i, (pos, pos_name) in enumerate(zip(positions, position_names)):
    color = cm(i / len(positions))
    ax_cell.plot(*pos, label=pos_name, marker="o", alpha=0.8, color=color, markersize=3)

if save_fig:
    fig_cell.savefig(figure_folder / "fig5_cell.pdf", transparent=True)
    
    for position in figs_traces:
        figs_traces[position].savefig(figure_folder / f"fig5_trace_{position}.pdf", transparent=True, bbox_inches='tight')
        figs_dists[position].savefig(figure_folder / f"fig5_dist_{position}.pdf", transparent=True, bbox_inches='tight')
